fix: move indexer methods to debug API
