+++ libraries/core_libs/network/rpc/Taraxa.h
@@ -29,9 +29,6 @@ class Taraxa : public TaraxaFace {
   virtual std::string taraxa_pbftBlockHashByPeriod(const std::string& _period) override;
   virtual Json::Value taraxa_getConfig() override;
   virtual Json::Value taraxa_getChainStats() override;
-  virtual Json::Value taraxa_getPeriodTransactionsWithReceipts(const std::string& _period) override;
-  virtual Json::Value taraxa_getPeriodDagBlocks(const std::string& _period) override;
-  virtual Json::Value taraxa_getPreviousBlockCertVotes(const std::string& _period) override;
 
  protected:
   std::weak_ptr<taraxa::FullNode> full_node_;
