+++ libraries/core_libs/network/rpc/TaraxaClient.h
@@ -106,33 +106,6 @@ class TaraxaClient : public jsonrpc::Client {
     else
       throw jsonrpc::JsonRpcException(jsonrpc::Errors::ERROR_CLIENT_INVALID_RESPONSE, result.toStyledString());
   }
-  Json::Value taraxa_getPeriodTransactionsWithReceipts(const std::string& param1) throw(jsonrpc::JsonRpcException) {
-    Json::Value p;
-    p.append(param1);
-    Json::Value result = this->CallMethod("taraxa_getPeriodTransactionsWithReceipts", p);
-    if (result.isObject())
-      return result;
-    else
-      throw jsonrpc::JsonRpcException(jsonrpc::Errors::ERROR_CLIENT_INVALID_RESPONSE, result.toStyledString());
-  }
-  Json::Value taraxa_getPeriodDagBlocks(const std::string& param1) throw(jsonrpc::JsonRpcException) {
-    Json::Value p;
-    p.append(param1);
-    Json::Value result = this->CallMethod("taraxa_getPeriodDagBlocks", p);
-    if (result.isObject())
-      return result;
-    else
-      throw jsonrpc::JsonRpcException(jsonrpc::Errors::ERROR_CLIENT_INVALID_RESPONSE, result.toStyledString());
-  }
-  Json::Value taraxa_getPreviousBlockCertVotes(const std::string& param1) throw(jsonrpc::JsonRpcException) {
-    Json::Value p;
-    p.append(param1);
-    Json::Value result = this->CallMethod("taraxa_getPreviousBlockCertVotes", p);
-    if (result.isObject())
-      return result;
-    else
-      throw jsonrpc::JsonRpcException(jsonrpc::Errors::ERROR_CLIENT_INVALID_RESPONSE, result.toStyledString());
-  }
 };
 
 }  // namespace net
